pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vex"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vex = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*", "stages*"]
